+++ sfmmultimodalapp-20250909/alibabacloud_sfmmultimodalapp20250909/models/_delete_memory_response_body.py
@@ -19,7 +19,6 @@ class DeleteMemoryResponseBody(DaraModel):
         self.data = data
         self.http_status_code = http_status_code
         self.message = message
-        # Id of the request
         self.request_id = request_id
         self.success = success