+++ sfmmultimodalapp-20250909/alibabacloud_sfmmultimodalapp20250909/models/_describe_mm_app_response_body.py
@@ -41,7 +41,6 @@ class DescribeMmAppResponseBody(DaraModel):
         self.modify_user_name = modify_user_name
         self.prompt = prompt
         self.publish_version = publish_version
-        # Id of the request
         self.request_id = request_id
         self.status = status