+++ sfmmultimodalapp-20250909/alibabacloud_sfmmultimodalapp20250909/models/_device_page_request.py
@@ -1,0 +1,57 @@
+# -*- coding: utf-8 -*-
+# This file is auto-generated, don't edit it. Thanks.
+from __future__ import annotations
+
+from darabonba.model import DaraModel
+
+class DevicePageRequest(DaraModel):
+    def __init__(
+        self,
+        app_id: str = None,
+        device_name: str = None,
+        page_number: int = None,
+        page_size: int = None,
+    ):
+        self.app_id = app_id
+        self.device_name = device_name
+        self.page_number = page_number
+        self.page_size = page_size
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.app_id is not None:
+            result['AppId'] = self.app_id
+
+        if self.device_name is not None:
+            result['DeviceName'] = self.device_name
+
+        if self.page_number is not None:
+            result['PageNumber'] = self.page_number
+
+        if self.page_size is not None:
+            result['PageSize'] = self.page_size
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AppId') is not None:
+            self.app_id = m.get('AppId')
+
+        if m.get('DeviceName') is not None:
+            self.device_name = m.get('DeviceName')
+
+        if m.get('PageNumber') is not None:
+            self.page_number = m.get('PageNumber')
+
+        if m.get('PageSize') is not None:
+            self.page_size = m.get('PageSize')
+
+        return self
+