+++ sfmmultimodalapp-20250909/alibabacloud_sfmmultimodalapp20250909/models/_update_command_request.py
@@ -2,7 +2,7 @@
 # This file is auto-generated, don't edit it. Thanks.
 from __future__ import annotations
 
-from typing import List
+from typing import List, Dict
 
 from alibabacloud_sfmmultimodalapp20250909 import models as main_models
 from darabonba.model import DaraModel
@@ -13,6 +13,7 @@ class UpdateCommandRequest(DaraModel):
         app_id: str = None,
         domain_code: str = None,
         domain_name: str = None,
+        reply_mode: str = None,
         tool_description: str = None,
         tool_examples: List[main_models.UpdateCommandRequestToolExamples] = None,
         tool_id: str = None,
@@ -24,6 +25,7 @@ class UpdateCommandRequest(DaraModel):
         self.app_id = app_id
         self.domain_code = domain_code
         self.domain_name = domain_name
+        self.reply_mode = reply_mode
         # This parameter is required.
         self.tool_description = tool_description
         self.tool_examples = tool_examples
@@ -57,6 +59,9 @@ class UpdateCommandRequest(DaraModel):
 
         if self.domain_name is not None:
             result['DomainName'] = self.domain_name
+
+        if self.reply_mode is not None:
+            result['ReplyMode'] = self.reply_mode
 
         if self.tool_description is not None:
             result['ToolDescription'] = self.tool_description
@@ -93,6 +98,9 @@ class UpdateCommandRequest(DaraModel):
         if m.get('DomainName') is not None:
             self.domain_name = m.get('DomainName')
 
+        if m.get('ReplyMode') is not None:
+            self.reply_mode = m.get('ReplyMode')
+
         if m.get('ToolDescription') is not None:
             self.tool_description = m.get('ToolDescription')
 
@@ -125,10 +133,14 @@ class UpdateCommandRequestToolParams(DaraModel):
         param_desc: str = None,
         param_example: str = None,
         param_name: str = None,
+        param_type: str = None,
+        required: bool = None,
     ):
         self.param_desc = param_desc
         self.param_example = param_example
         self.param_name = param_name
+        self.param_type = param_type
+        self.required = required
 
     def validate(self):
         pass
@@ -147,6 +159,12 @@ class UpdateCommandRequestToolParams(DaraModel):
         if self.param_name is not None:
             result['ParamName'] = self.param_name
 
+        if self.param_type is not None:
+            result['ParamType'] = self.param_type
+
+        if self.required is not None:
+            result['Required'] = self.required
+
         return result
 
     def from_map(self, m: dict = None):
@@ -160,13 +178,21 @@ class UpdateCommandRequestToolParams(DaraModel):
         if m.get('ParamName') is not None:
             self.param_name = m.get('ParamName')
 
+        if m.get('ParamType') is not None:
+            self.param_type = m.get('ParamType')
+
+        if m.get('Required') is not None:
+            self.required = m.get('Required')
+
         return self
 
 class UpdateCommandRequestToolExamples(DaraModel):
     def __init__(
         self,
+        parameters: Dict[str, str] = None,
         query: str = None,
     ):
+        self.parameters = parameters
         self.query = query
 
     def validate(self):
@@ -177,6 +203,9 @@ class UpdateCommandRequestToolExamples(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.parameters is not None:
+            result['Parameters'] = self.parameters
+
         if self.query is not None:
             result['Query'] = self.query
 
@@ -184,6 +213,9 @@ class UpdateCommandRequestToolExamples(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('Parameters') is not None:
+            self.parameters = m.get('Parameters')
+
         if m.get('Query') is not None:
             self.query = m.get('Query')