+++ sfmmultimodalapp-20250909/ChangeLog.md
@@ -1,3 +1,25 @@
+2026-06-11 Version: 1.3.0
+- Support API DevicePage.
+- Support API DeviceUpdate.
+- Support API QueryAppQuota.
+- Update API CreateCommand: add request parameters ReplyMode.
+- Update API CreateCommand: add request parameters ToolExamples.$.Parameters.
+- Update API CreateCommand: add request parameters ToolParams.$.ParamType.
+- Update API CreateCommand: add request parameters ToolParams.$.Required.
+- Update API DescribeCommand: add response parameters Body.ReplyMode.
+- Update API DescribeCommand: add response parameters Body.ToolExamples.$.Parameters.
+- Update API DescribeCommand: add response parameters Body.ToolParams.$.ParamType.
+- Update API DescribeCommand: add response parameters Body.ToolParams.$.Required.
+- Update API ListCommand: add response parameters Body.ToolInfoList.$.ReplyMode.
+- Update API ListCommand: add response parameters Body.ToolInfoList.$.ToolExamples.$.Parameters.
+- Update API ListCommand: add response parameters Body.ToolInfoList.$.ToolParams.$.ParamType.
+- Update API ListCommand: add response parameters Body.ToolInfoList.$.ToolParams.$.Required.
+- Update API UpdateCommand: add request parameters ReplyMode.
+- Update API UpdateCommand: add request parameters ToolExamples.$.Parameters.
+- Update API UpdateCommand: add request parameters ToolParams.$.ParamType.
+- Update API UpdateCommand: add request parameters ToolParams.$.Required.
+
+
 2026-04-20 Version: 1.2.0
 - Support API MmAppBindingMcp.
 - Support API MmAppBindingRag.